Add run-time configuration and configuration files
In order to make DoCitten more configurable and re-usable, a number
of options and settings are now being made configurable at run-time.
These changes also partially address #1, as loading a Service is now
a self-contained operation.

We add a very simple scripting language to DoCitten, currently
containing the following verbs:

 - "#"    -- Used to denote a comment line in a script
 - "log"  -- Write a string to the log
 - "loglevel" -- Change the logging level
 - "nick" -- Set or change the Bot's nickname
 - "connect" -- Connect to an IRC server
 - "join" -- Join one or more channels
 - "part" -- Leave one or more channels
 - "load" -- Load a Service
 - "quit" -- Performs a clean shutdown and exits

The verb "unload" is also reserved, but is not yet implemented.
The verb "nick" is not fully implemented, as Mewler does not
support changing nick after connection.

The old content of Main.java has been rewritten as docittenrc.sample.

DoCitten will search for config files in:

 - The first argument to main()
 - "docittenrc" in the current directory
 - "/etc/docittenrc"

If no configuration files are available, the system exits.

Once the configuration file has been parsed without a break
(currently caused only be a "quit" command or an IOException),
command input is parsed from the standard input stream.