Fix for mismatched timezones in CatPicOfDayService
The lastUpdated and Now calendar objects were created with
different timezones, causing the date to be considered as
'changed' for the entire period the two calendars were out
of sync.

For each of use, a second constructor is added which allows
the user to explictly specify the TimeZone to use.