Switch out HashTable for HashMap
From http://docs.oracle.com/javase/7/docs/api/java/util/Hashtable.html:

"As of the Java 2 platform v1.2, this class was retrofitted to implement
the Map interface, making it a member of the Java Collections Framework.
Unlike the new collection implementations, Hashtable is synchronized. If a
thread-safe implementation is not needed, it is recommended to use HashMap
in place of Hashtable. If a thread-safe highly-concurrent implementation
is desired, then it is recommended to use ConcurrentHashMap in place of
Hashtable."

In this case, I have replaced 'tracking' with with a ConcurrentHashMap.