Move build-time Java version check to dedicated script
The build time test that the current Java version is installed
is not a specific makefile target, that leaves behind a blank
file as an artefact (additionally allowing a new method for people
to override the check).
The check is now performed by a shell script, which has a slightly
higher level of complexity in the what version strings it will accept.